Data Engineering_MS2/DE_MS2_MET_P1_52_24625/requirements.txt
numpy
pandas
scikit-learn
requests
beautifulsoup4
PyYAML
SQLAlchemy
psycopg2-binary
kafka-python
docker